refactor : 도메인별 패키지 분리
