+++ src/main/java/step3/domain/amount/PurchasingAmount.java
@@ -1,4 +1,4 @@
-package step3.domain;
+package step3.domain.amount;
 
 public class PurchasingAmount {
