+++ src/test/java/step3/domain/amount/AmountTest.java
@@ -1,4 +1,4 @@
-package step3.domain;
+package step3.domain.amount;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
