+++ src/test/java/step3/domain/input/InputPurchasingAmountTest.java
@@ -4,8 +4,8 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import step3.domain.Amount;
-import step3.domain.PurchasingAmount;
+import step3.domain.amount.Amount;
+import step3.domain.amount.PurchasingAmount;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
